requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pandas
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
